Added gettimeofday() system call and C lib functions

+++ kernel/includes/syscall.h
@@ -9,6 +9,6 @@
 #include <hw/types.h>
 
 extern "C" int _syscall(void);
-int syscall(int nr, UInt64 arg0, UInt64 arg1, UInt64 arg2, UInt64 arg3, UInt64 arg4);
+UInt64 syscall(int nr, UInt64 arg0, UInt64 arg1, UInt64 arg2, UInt64 arg3, UInt64 arg4);
 
 #endif // SYSCALL_H
